Check for network checkpoint on page load
When using IP addresses to access the web UI and changing the network,
it is not possible to clear the network checkpoint because the IP
changes. In order to be able to clear the checkpoint in this scenario,
the web UI will check for the existence of a checkpoint during page
load and show a dialog if one exists to keep or rollback the
network changes.