Handle WebRTC signaling exceptions
We can detect unsupported video codecs and provide user feedback about that
fact by observing errors on RTCPeerConnection. The errors are throwns when we
set the remote session description or adding a ICE candidate.

Firefox closes the transceiver for unsupported video tracks. We track this as
well and inform the user with an error message.